Big refactor, add Client, split out packets to own file
This adds a Client class that merges a bunch of the funcitonality of the
bleak client and rx_handler (which should have been tx handler...) into
a single object/class that should be easier to work with.

I also split out each packet and parser into their own files, so that if
someone wants to understand the packets they can ignore the python/bleak
specific stuff.

The only issue was that the "get_realtime_heart_rate" style commands
which may send multiple packets or do multiple things with the response
now live on the client. Maybe I can find another way to structure this
so the client is more agnostic, but I think this is OK. It does spread
the context for how a feature works across 2 files, but oh well.