fix: don't trunate hr log data on parse
Fix a bug where the minutes_so_far mixed timezones up so that data was
dropped when it shold not be

I still don't like the fact that parsing is date dependent, I think it
might be possible to remove the is_today check and just use the index
instead. Future readings will be 0 should be ignored.