feat: add sync command to write data to sqlite
Introduce sqlalchemy as a dep so that we can store and sync data from
the ring over time. This add a db module that contains ORM models and
functions to write data to the database

I also got sick of all the date stuff so I added date_utils as a module
to keep everything in one place.